requirements.txt
spotipy
pandas
numpy
scikit-learn
matplotlib
seaborn
gradio
python-dotenv
joblib
requests
jupyter
plotly